+++ CHANGELOG.md
@@ -1,5 +1,15 @@
 # Changelog
 
+## 1.2.0
+
+- add helper method: setDate as milliseconds on pikaday 
+- add helper method: setMinDate as milliseconds on pikaday 
+- add helper method: setMaxDate as milliseconds on pikaday 
+
+## 1.1.0
+
+- add helper method: get milliseconds from pikaday
+
 ## 1.0.0
 
 - Initial version